Refactor CheckIfTransactionsExist/sumTransactions, fix formatting bug  in ToDollars
code review - CheckIfTransactionsExist(bool checkAll) - this is badly named for what it does, and why feed in a bool if you do nothing with it? Should rename to tell what it does - but it actually fits entirely within the calling method at the moment, so shift code upwards.  Even for the expansion/future state, summing values here is bad coding.  Check should check.